Refactor file upload handling to include file extension validation in read_file utility

+++ api/utils/utils.py
@@ -1,10 +1,20 @@
 from pdfminer.high_level import extract_text
 import os
 
-def read_file(content:bytes):
-    with open("temp.pdf", "wb") as temp_file:
-        temp_file.write(content)
 
-    file_content = extract_text("temp.pdf")
-    os.remove(path=os.getcwd() + "\\temp.pdf")
-    return file_content
+def read_file(content: bytes, extension: str):
+    if extension== "pdf":
+        with open("temp.pdf", "wb") as temp_file:
+            temp_file.write(content)
+
+        file_content = extract_text("temp.pdf")
+        os.remove(path=os.getcwd() + "\\temp.pdf")
+        return file_content
+
+    elif extension == "txt":
+        return content.decode()
+
+    else:
+        return "Invalid file format!"
+
+
